Do not use: sub-module to prevent date change on RefreshLess prefetch:
The problem with this is that this results in a prefetch never changing
the date, even if it was used for a real navigation, because the
response displayed to the user will at best lag one response behind and
in a lot of cases won't update at all. There are only some edge cases
where incorrectly updating the date due to a prefetch are a
(minor) problem because it's always updated at the start of the
request, whether used or not.