WIP attempt at using a header to set current date; not working:
The intent is to try to pass the actual date to the back-end when
navigating to a page that isn't a wiki node (which sets the date). For
example, the random route and any non-wiki node will use whatever the
last prefetch or preload would set as the current date, which is a
problem.