WIP RefreshLess current date save and restore; not working:
The assumption here is that the current date is matches the current page
the user is looking at before it's changed, and to restore that at the
end of the request. This assumption is probably wrong, however, because
it doesn't seem to quite work and still results in incorrect dates.